% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.r
\name{batch}
\alias{batch}
\title{Batch Execution}
\usage{
batch(addr = "localhost", port = 55555, file, timer = FALSE)
}
\arguments{
\item{addr}{The remote host/address/endpoint.}

\item{port}{The port (number) that will be used for communication between 
the client and server.  The port value for the client and server
must agree.}

\item{file}{A character string pointing to the file you wish to execute/source.}

\item{timer}{Logical; should the "performance prompt", which shows timing
statistics after every command, be used?}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
Run a script on a remote server in batch.  Similar to R's own
\code{source()} function.
}

