% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_client_utils.r
\name{evalc}
\alias{evalc}
\title{evalc}
\usage{
evalc(expr)
}
\arguments{
\item{expr}{Expression to be evaluated on the client.}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
A function to evaluate expressions on the client's R session.  To
eval expressions on the server, just use \code{eval()}.  Instead of
using this function, you could also just kill the client, do your
local operations, then re-run your \code{client()} command.
}
\details{
Evaluate expressions on the client.
}
