% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.r
\name{has.sodium}
\alias{has.sodium}
\title{has.sodium}
\usage{
has.sodium()
}
\value{
Returns \code{TRUE} if the sodium package is available, and 
\code{FALSE} otherwise.
}
\description{
Report if the sodium package is availabe for use.
}
