% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_internals.r, R/comms.r, R/logging.r, R/server.r, R/state.r, R/utils.r
\name{internals}
\alias{internals}
\alias{logprint}
\alias{remoter_error}
\alias{remoter_receive}
\alias{remoter_send}
\alias{remoter_warning}
\alias{reset_state}
\alias{validate_port}
\title{Remoter Internals}
\usage{
remoter_send(data, send.more = FALSE)

remoter_receive()

logprint(msg, checkverbose = FALSE, checkshowmsg = FALSE, preprint = "",
  level = "")

remoter_warning(warn)

remoter_error(err)

reset_state()

validate_port(port)
}
\description{
Internal remoter utilities needed to create a custom server backend
for use with the remoter client.  See pbdCS for an example.
}

