% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.r
\name{is.secure}
\alias{is.secure}
\title{is.secure}
\usage{
is.secure()
}
\value{
Returns \code{TRUE} if messages between client and server are
currently encrypted, and \code{FALSE} if not.  If the client 
is not currently running (i.e., if executed from just a regular
R prompt), then \code{NA} is returned.
}
\description{
Report if communications with the connected server are
encrypted.
}

