% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbd_launch_servers}
\alias{pbd_launch_client}
\alias{pbd_launch_servers}
\title{Client/Server Launchers}
\usage{
pbd_launch_servers(nranks = 2, bcast_method = "zmq", port = 5555)

pbd_launch_client(port = 5555)
}
\arguments{
\item{nranks}{The number of MPI ranks to launch.}

\item{bcast_method}{The method used by the servers to communicate.  Options are "zmq"
for ZeroMQ-based communication, or "mpi" for}

\item{port}{The port to use for communication between the client and rank 0.}

\item{intern}{logical; determines if the output of the evaluation of the
script should be saved as an R character vector.}
}
\description{
This function is a simple wrapper around the system() command.  As such,
data is not shared between the calling R process and the batch processes
which execute the 'body' source.
}
\details{
Launchers for the pbdR client/server.

TODO: describe structure
}

