% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{ls.local}
\alias{ls.local}
\title{ls.local}
\usage{
ls.local(envir, all.names = FALSE, pattern)
}
\arguments{
\item{envir}{Environment (as in \code{ls()}).}

\item{all.names}{Logical that determines if all names are returned or those beginning
with a '.' are omitted (as in \code{ls()}).}

\item{pattern}{Optional regular expression (as in \code{ls()}).}
}
\description{
A function to view environments on the client's R session.  To
view objects on the server, use \code{ls()}.
}
\details{
View objects on the client.
}

