% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_client_utils.r
\name{lsc}
\alias{lsc}
\title{ls on Client}
\usage{
lsc(envir, all.names = FALSE, pattern)
}
\arguments{
\item{envir}{Environment (as in \code{ls()}).}

\item{all.names}{Logical that determines if all names are returned or those beginning
with a '.' are omitted (as in \code{ls()}).}

\item{pattern}{Optional regular expression (as in \code{ls()}).}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
A function to view environments on the client's R session.  To
view objects on the server, just use \code{ls()}.  Instead of
using this function, you could also just kill the client, do your
local operations, then re-run your \code{client()} command.
}
\details{
View objects on the client.
}

