% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.r
\name{password}
\alias{password}
\title{Password Input}
\usage{
password()
}
\description{
Masked user input (where supported; see details section) for
entering passwords.
}
\details{
Masking (i.e., not displaying the password or "*" characters
as input is provided) is supported on sevreal, but not all
platforms.  It is supported in RStudio, provided you
have a suitable version of the GUI and of the package 
'rstudioapi'.  It should also work in the terminal on any
OS.  It is *NOT* supported in the Windows GUI RGui or on
the Mac GUI R.app.
}

