% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{pbd_exit}
\alias{pbd_exit}
\title{pbd_exit}
\usage{
pbd_exit(clientonly = FALSE)
}
\description{
This function cleanly shuts down any pbdR servers which have been
spawned, as well as shutting down the client.  Failing to use
this function to shut down the client/server may cause unexpected
results.
}
\details{
Exit the pbdR client/server.
}

