% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbd_repl}
\alias{pbd_repl}
\title{pbd_repl}
\usage{
pbd_repl(env = sys.parent())
}
\arguments{
\item{env}{Environment where repl evaluations will take place.}
}
\description{
This is exported for clean access reasons; you shoud not directly
use this function.
}
\details{
The REPL for the client/server.
}

