% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cs-package.r
\docType{package}
\name{pbdCS-package}
\alias{pbdCS-package}
\title{pbdR Client/Server Utilities}
\description{
pbdR Client/Server Utilities
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\keyword{package}

