% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\docType{data}
\name{pbdenv}
\alias{pbdenv}
\title{State management for the pbdR Client/Server}
\format{\preformatted{<environment: 0x2e33f98> 
}}
\usage{
pbdenv
}
\description{
State management for the pbdR Client/Server
}
\keyword{datasets}

