% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recvfile.r
\name{recvfile}
\alias{recvfile}
\title{recvfile}
\usage{
recvfile(file_send, file_recv, verbose = TRUE)
}
\arguments{
\item{file_recv, file_send}{The file paths (as strings) for the input/sent file and the output/received
file.}

\item{verbose}{Should file transfer information be printed?}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
Transfer file from server to client.
}
