% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relay.r
\name{relay}
\alias{relay}
\title{Relay Launcher}
\usage{
relay(addr, recvport = 55556, sendport = 55555, verbose = FALSE)
}
\arguments{
\item{addr}{The address of the server.}

\item{recvport}{The port for receiving commands from the client.}

\item{sendport}{The port for sending commands to the server.}

\item{verbose}{Show verbose messaging.}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
Launcher for the remoter relay.
}
\details{
The relay is an intermediary or "middleman" between the client
and server meant for machines with split login/compute nodes.
}
