% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoter-package.r
\docType{package}
\name{remoter-package}
\alias{remoter-package}
\title{remoter}
\description{
A package for running R commands on a remote R session.  With 
remoter, you can use RStudio running on your laptop to execute
commands on an R session running on, for example, Amazon's EC2.
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\references{
Project URL: \url{https://github.com/RBigData/remoter}
}
\keyword{package}

