% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{remoter_localize}
\alias{remoter_localize}
\title{remoter_localize}
\usage{
remoter_localize(object, newname, env = .GlobalEnv)
}
\arguments{
\item{object}{A remote R object.}

\item{newname}{The name the object should take when it becomes local. If left blank,
the local name will have the original (remote) object's name.}

\item{env}{The environment into which the assignment will take place. The
default is the global environment.}
}
\description{
This function allows you to pass an object from MPI rank 0 of
the servers to the local R session behind the client.
}
\details{
Localize R objects.
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(pbdCS)
> y
###  Error: object 'y' not found
> remoter_launch_servers()
> remoter_launch_client()
pbdR> x
### Error: object 'x' not found
pbdR> x <- "some data"
pbdR> x
###  [1] "some data"
pbdR> remoter_localize(x, "y")
pbdR> remoter_exit()
> y
###  [1] "some data"
}
}

