% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{remoter_repl}
\alias{remoter_repl}
\title{remoter_repl}
\usage{
remoter_repl(env = sys.parent())
}
\arguments{
\item{env}{Environment where repl evaluations will take place.}
}
\description{
This is exported for clean access reasons; you shoud not directly
use this function.
}
\details{
The REPL for the client/server.
}

