% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_client_utils.r
\name{rmc}
\alias{rmc}
\title{rmc}
\usage{
rmc(..., list = character(), envir)
}
\arguments{
\item{...}{Objects to be removed from the client's R session.}

\item{list}{Character vector naming objects to be removed (as in \code{rm()}).}

\item{envir}{Environment (as in \code{rm()}).}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
A function to remove objects from the client's R session.  To
remove objects on the server, just use \code{rm()}.  Instead of
using this function, you could also just kill the client, do your
local operations, then re-run your \code{client()} command.
}
\details{
Remove objects on the client.
}

