% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_dev_rrpng.r
\name{rrpng}
\alias{rrpng}
\title{rrpng}
\usage{
rrpng(func = NULL, filename = tempfile(fileext = "_rr.png"), width = 480,
  height = 480, units = "px", pointsize = 12, bg = "white", res = NA,
  ...)
}
\arguments{
\item{func}{A function generates a plot}

\item{filename}{A temporary file to save the plot on server}

\item{width}{width of the plot as in \code{grDevices::png()}}

\item{height}{height of the plot as in \code{grDevices::png()}}

\item{units}{units of the width and height as in \code{grDevices::png()}}

\item{pointsize}{pointsze of the plotted text as in \code{grDevices::png()}}

\item{bg}{background colour as in \code{grDevices::png()}}

\item{res}{resolution as in \code{grDevices::png()}}
}
\value{
A local plot
}
\description{
Provide a graphic device locally for plots generated on server of Remote R
}
\details{
Remote R PNG Device
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(remoter)
> client()
remoter> a <- function() plot(1:5)
remoter> rrpng(a)
remoter> dev.newc()
remoter> b <- function() plot(iris$Sepal.Length, iris$Petal.Length)
remoter> rrpng(b)
remoter> dev.offc(2)
remoter> dev.offc(3)
remoter> q()
}

}

