% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendfile.r
\name{sendfile}
\alias{sendfile}
\title{sendfile}
\usage{
sendfile(file_send, file_recv, verbose = TRUE)
}
\arguments{
\item{file_send, file_recv}{The file paths (as strings) for the input/sent file and the output/received
file.}

\item{verbose}{Should file transfer information be printed?}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
Transfer file from client to server.
}
