% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{eval.local}
\alias{eval.local}
\title{eval.local}
\usage{
eval.local(expr)
}
\arguments{
\item{expr}{Expression to be evaluated on the client.}
}
\description{
A function to evaluate expressions on the client's R session.
}
\details{
Evaluate expressions on the client.
}

