% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit.r
\name{exit}
\alias{exit}
\title{exit}
\usage{
exit(client.only = TRUE)
}
\arguments{
\item{client.only}{Logical; if \code{TRUE}, then the client disconnects from
the server.  Otherwise, the server is shut down together 
with the client.}
}
\value{
Returns \code{TRUE} invisibly on successful exit.
}
\description{
This function cleanly shuts down the remoter server the client
is currently connected to, as well as shutting down the client.
One can also use \code{q()} (while the client is running), and
this will not close the active R session on the client.
}
\details{
Exit the remoter client/server.
}

