% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launchers.r
\name{pbd_launch_servers}
\alias{pbd_launch_client}
\alias{pbd_launch_servers}
\title{Client/Server Launchers}
\usage{
pbd_launch_servers(nranks = 2, bcast_method = "zmq", port = 5555)

pbd_launch_client(port = 5555)
}
\arguments{
\item{nranks}{The number of MPI ranks to launch.}

\item{bcast_method}{The method used by the servers to communicate.  Options are "zmq"
for ZeroMQ-based communication, or "mpi" for}

\item{port}{The port to use for communication between the client and rank 0.}
}
\description{
This function is a simple wrapper around the system() command.  As such,
data is not shared between the calling R process and the batch processes
which execute the 'body' source.
}
\details{
Launchers for the pbdR client/server.

The \code{port} values between the client and server \emph{MUST}
agree.  If they do not, this can cause the client to hang.

The servers are launched via \code{pbdRscript()}.

The client is a specialized REPL that intercepts commands sent
through the R interpreter.  These commands are then sent from the
client to and evaluated on the servers.

The client communicates over ZeroMQ with MPI rank 0 (of the
servers) using a REQ/REP pattern.  Both commands (from client to
server) and returns (from servers to client) are handled in this
way.  Once a command is sent from the client to MPI rank 0,
that command is then "broadcasted" from MPI rank 0 to the other
MPI ranks.  The method of broadcast is handled by the input
\code{bcast_method}.  If \code{bcast_method="mpi"}, then \code{MPI_bcast}
is used to transmit the command.  Otherwise (\code{bcast_method="zmq"})
uses ZeroMQ with a PUSH/PULL pattern.  The MPI method is probably
epsilon faster, but it will busy-wait.  The ZeroMQ bcast method
will not busy wait, in addition to the other benefits ZeroMQ
affords; thus, \code{bcast_method="zmq"} is the default.

To shut down the servers and the client, use the command \code{pbd_exit()}.
}
\examples{
\dontrun{
library(pbdCS)
pbd_launch_servers()
pbd_launch_client()
}
}
\seealso{
\code{\link{pbdRscript}, \link{pbd_exit}}
}

