% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{pbd_localize}
\alias{pbd_localize}
\title{pbd_localize}
\usage{
pbd_localize(object, newname)
}
\arguments{
\item{object}{}

\item{newname}{}
}
\description{
This function allows you to pass an object from MPI rank 0 of
the servers to the local R session behind the client.
}
\details{
Localize R objects.
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(pbdCS)
> y
###  Error: object 'y' not found
> pbd_launch_servers()
> pbd_launch_client()
pbdR> x
### Error: object 'x' not found
pbdR> x <- "some data"
pbdR> x
###  [1] "some data"
pbdR> pbd_localize(x, "y")
pbdR> pbd_exit()
> y
###  [1] "some data"
}
}

