% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_rpng.r, R/wcc_rpng_utils.r
\name{rpng}
\alias{dev.off}
\alias{rpng}
\alias{rpng.new}
\alias{rpng.off}
\title{rpng}
\usage{
rpng.new(expr, filename = NULL, width = 590, height = 590, units = "px",
  pointsize = 12, bg = "white", res = 120, interpolate = TRUE, ...)

rpng.off(which = grDevices::dev.cur())

dev.off(which = grDevices::dev.cur())
}
\arguments{
\item{expr}{An expression or a function generating a plot. This checks in the
following orders: expression or ggplot. The ggplot
are eval'd within the \code{rpng.new()}, while the expression is
eval'd at \code{parent.frame()}.}

\item{filename}{A temporary file to save the plot on server}

\item{width}{width of the plot as in \code{grDevices::png()}}

\item{height}{height of the plot as in \code{grDevices::png()}}

\item{units}{units of the width and height as in \code{grDevices::png()}}

\item{pointsize}{pointsze of the plotted text as in \code{grDevices::png()}}

\item{bg}{background colour as in \code{grDevices::png()}}

\item{res}{resolution as in \code{grDevices::png()}}

\item{interpolate}{if render the image with interpolation as in
\code{grDevices::rasterImage()}. Default is changed to FALSE. It gains a few
readiness in low res or small images when this argument is TRUE.}

\item{...}{additional arguments as in \code{grDevices::png()}}

\item{which}{An integer specifying a device number as in \code{grDevices::dev.off()}}
}
\description{
Provide a graphic device locally for plots generated on server of Remote R

\code{rpng()} generates locally a device/window.

\code{rpng.new()} generates locally a device/window.

\code{rpng.off()} turns off locally a device/window.

\code{dev.off()} is an alias of \code{rpng.off()} in order to consisten
with th original device function \code{grDevices::dev.off()}.
}
\details{
Remote R PNG Device
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs
### remotely
> library(remoter, quietly = TRUE)
> client()

remoter> plot(1:5)
remoter> dev.off()

remoter> rpng()
remoter> plot(iris$Sepal.Length, iris$Petal.Length)
remoter> dev.off()

remoter> library(ggplot2)
remoter> g1 <- ggplot(iris, aes(x = Sepal.Length, y = Petal.Length,
remoter+              color = Species)) +
remoter+       geom_point(aes(shape = Species))
remoter> rpng()
remoter> print(g1)
remoter> dev.off()

remoter> g2 <- g1 + geom_smooth(method = "lm")
remoter> rpng(bg = "transparent")
remoter> g2
remoter> dev.off()

remoter> rpng.new(plot(1:5))

remoter> rpng.new(g1)

remoter> b <- function() plot(iris$Sepal.Length, iris$Petal.Length)
remoter> rpng.new(b)

remoter> q()
>
}

}
\seealso{
\code{\link{rDevices}}
}

