% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoter-package.r
\docType{package}
\name{remoter-package}
\alias{remoter-package}
\title{remoter}
\description{
A set of utilities for client/server computing with R, controlling
a remote R session (the server) from a local one (the client).  Simply set
up a server (see package vignette for more details) and connect to it from
your local R session ('RStudio', terminal, etc).  The client/server
framework is a custom 'REPL' and runs entirely in your R session without the
need for installing a custom environment on your system.  Network
communication is handled by the 'ZeroMQ' library by way of the 'pbdZMQ'
package.
}
\references{
Project URL: \url{https://github.com/RBigData/remoter}
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\keyword{package}
