% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repl.r
\name{rm.local}
\alias{rm.local}
\title{rm.local}
\usage{
rm.local(..., list = character(), envir)
}
\arguments{
\item{...}{Objects to be removed from the client's R session.}

\item{list}{Character vector naming objects to be removed (as in \code{rm()}).}

\item{envir}{Environment (as in \code{rm()}).}
}
\description{
A function to remove objects from the client's R session.  To
remove objects on the server, use \code{rm()}.
}
\details{
View objects on the client.
}

