% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_dev.r
\name{dev.offc}
\alias{dev.offc}
\title{rrDevices}
\usage{
dev.offc(which = dev.cur())
}
\arguments{
\item{which}{An integer specifying a device number as in \code{grDevices::dev.off()}}

\item{...}{arguments to be passed to the device selected as in
\code{grDevices::dev.new()}}
}
\description{
Functions for controlling graphic device locally

\code{dev.newc()} locally eval \code{grDevices::dev.new()}.

\code{dev.offc()} locally evals \code{grDevices::dev.off()}.

\code{dev.curc()} locally evals \code{grDevices::dev.cur()}.
}
\details{
Remote R Graphic Device Controling Functions
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs remotely
> library(remoter)
> client()
remoter> a <- function() plot(1:5)
remoter> rrpng(a)
remoter> dev.newc()
remoter> b <- function() plot(iris$Sepal.Length, iris$Petal.Length)
remoter> rrpng(b)
remoter> dev.curc()
remoter> dev.offc(2)
remoter> dev.offc(3)
remoter> q()
}

}

