% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcc_dev_rrpng.r
\name{rrpng}
\alias{rrpng}
\title{rrpng}
\usage{
rrpng(expr, filename = tempfile(fileext = "_rr.png"), width = 590,
  height = 590, units = "px", pointsize = 12, bg = "white", res = 120,
  interpolate = FALSE, ...)
}
\arguments{
\item{expr}{An expression or a function generating a plot. This checks in the
following orders: function, ggplot, expression. The function and ggplot
are eval'd within the \code{rrpng()}, while the expression is
eval'd at \code{parent.frame()}.}

\item{filename}{A temporary file to save the plot on server}

\item{width}{width of the plot as in \code{grDevices::png()}}

\item{height}{height of the plot as in \code{grDevices::png()}}

\item{units}{units of the width and height as in \code{grDevices::png()}}

\item{pointsize}{pointsze of the plotted text as in \code{grDevices::png()}}

\item{bg}{background colour as in \code{grDevices::png()}}

\item{res}{resolution as in \code{grDevices::png()}}

\item{interpolate}{if render the image with interpolation as in
\code{grDevices::rasterImage()}. Default is changed to FALSE. It gains a few
readiness in low res or small images when this argument is TRUE.}

\item{...}{additional arguments as in \code{grDevices::png()}}
}
\value{
A local plot
}
\description{
Provide a graphic device locally for plots generated on server of Remote R
}
\details{
Remote R PNG Device
}
\examples{
\dontrun{
### Prompts are listed to clarify when something is eval'd locally vs
### remotely
> library(remoter, quietly = TRUE)
> client()

remoter> rrpng(plot(1:5))
remoter> b <- function() plot(iris$Sepal.Length, iris$Petal.Length)
remoter> rrpng(b, bg = "transparent")

remoter> library(ggplot2)
remoter> g <- ggplot(iris, aes(x = Sepal.Length, y = Petal.Length)) +
remoter+      geom_point(aes(colour = Species, shape = Species))
remoter> rrpng(g)
remoter> dev.newc(width = 6, height = 4)
remoter> rrpng(g, width = 6 * 72, height = 4 * 72, bg = "transparent")

remoter> q()
>
}

}
\seealso{
\code{\link{rrDevices}}
}

