% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.r
\name{server}
\alias{server}
\title{Server Launcher}
\usage{
server(port = 55555, password = NULL, maxretry = 5, showmsg = FALSE)
}
\arguments{
\item{port}{The port (number) that will be used for communication between
the client and server.  The port value for the client and server
must agree.}

\item{password}{A password the client must enter before the user can process
commands on the server.}

\item{maxretry}{The maximum number of retries for passwords before shutting
everything down.}

\item{showmsg}{Logical; if TRUE, turns on the "debug mode" for the server,
and prints messages in the server terminal.}
}
\description{
Launcher for the remoter server.
}

